% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_get.R
\name{convert_column_names}
\alias{convert_column_names}
\title{Convert Column Names to snake_case}
\usage{
convert_column_names(df)
}
\arguments{
\item{df}{A data frame whose column names are to be converted.}
}
\value{
The data frame with its column names converted to snake_case.
}
\description{
Convert Column Names to snake_case
}
\keyword{internal}
