% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_get.R
\name{convert_column_types}
\alias{convert_column_types}
\title{Convert Column Types in a Data Frame}
\usage{
convert_column_types(df)
}
\arguments{
\item{df}{A data frame containing columns to be processed.}
}
\value{
A data frame with updated column types.
}
\description{
Convert Column Types in a Data Frame
}
\keyword{internal}
