% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmp_set_api_key.R
\name{fmp_set_api_key}
\alias{fmp_set_api_key}
\title{Set the Financial Modeling Prep API key}
\usage{
fmp_set_api_key()
}
\value{
Invisibly returns \code{TRUE} after saving the key or aborting the
operation. The function will read the \code{.Renviron} file after saving,
allowing the environment variables to be immediately available. A restart of
the R session is recommended.
}
\description{
This function prompts the user to input their Financial Modeling Prep (FMP)
API key and saves it to a \code{.Renviron} file, either in the project directory
or the home directory. It also gives the user the option to add \code{.Renviron}
to \code{.gitignore} for security purposes.
}
