% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmpapi-package.R
\docType{package}
\name{fmpapi-package}
\alias{fmpapi}
\alias{fmpapi-package}
\title{fmpapi: Flexible & Tidy Client for the Financial Modeling Prep API}
\description{
Client for the Financial Modeling Prep (FMP) API \url{https://site.financialmodelingprep.com/developer/docs} that supports all available resources and returns data in a tidy data format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-finance/r-fmpapi}
  \item Report bugs at \url{https://github.com/tidy-finance/r-fmpapi/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID}) [copyright holder]

}
\keyword{internal}
