% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_cap_historical.R
\name{get_market_cap_historical}
\alias{get_market_cap_historical}
\title{Get Historical Market Capitalization}
\usage{
get_market_cap_historical(symbol, start_date, end_date)
}
\arguments{
\item{symbol}{Character. The stock ticker symbol of the company for which to retrieve market capitalization data.}

\item{start_date}{Date. The start date for the data retrieval in "YYYY-MM-DD" format.}

\item{end_date}{Date. The end date for the data retrieval in "YYYY-MM-DD" format.}
}
\value{
A tibble with historical market capitalization data, arranged by date.
}
\description{
Retrieves historical market capitalization data for a specified symbol
within a defined date range.
}
\details{
The function queries an API to obtain market capitalization data for the specified `symbol` between `start_date` and `end_date`.
The data is returned as a tibble with appropriately converted column names and types.
}
\examples{
\dontrun{
get_market_cap_historical("AAPL", "2020-01-01", "2020-12-31")
}

}
