% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_column_names.R
\name{convert_column_names}
\alias{convert_column_names}
\title{Convert column names to snake_case}
\usage{
convert_column_names(df)
}
\arguments{
\item{df}{A data frame whose column names are to be converted.}
}
\value{
The data frame with its column names converted to snake_case.
}
\description{
This function converts the column names of a data frame from camelCase to snake_case by
inserting underscores between lowercase and uppercase letters and converting all letters to lowercase.
}
\keyword{internal}
