% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmpapi-package.R
\docType{package}
\name{fmpapi-package}
\alias{fmpapi}
\alias{fmpapi-package}
\title{fmpapi: Tidy Financial Modeling Prep API Client}
\description{
Tidy client for the Financial Modeling Prep (FMP) API \url{https://site.financialmodelingprep.com/developer/docs}. This package allows users to easily fetch financial data such as company profiles, balance sheet statements, income statements, and cash flow statements, using the FMP API. Functions return tidy data frames with snake_case column names, making it easy to integrate the results into common workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-finance/r-fmpapi}
  \item Report bugs at \url{https://github.com/tidy-finance/r-fmpapi/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID}) [copyright holder]

}
\keyword{internal}
